Rename postreconcile to commit.
